==> notebooks/00_merge_features_sample.ipynb <==
%matplotlib inline

import os
import sys
import glob
import pandas as pd

features_dir = '../data/features/samples/'
features_all = glob.glob(os.path.join(features_dir, '*all.csv'))

dfs = []

for path in features_all:
    stem = path[:path.find('-all')]
    noise_path = stem + '-noise.csv'
    petrel_path = stem + '-petrel.csv'
    all_df = pd.read_csv(path)
    noise_df = pd.read_csv(noise_path)
    petrel_df = pd.read_csv(petrel_path)
    all_df_noise = all_df.loc[noise_df['id']].copy()
    all_df_noise['petrel'] = 0
    all_df_petrel = all_df.loc[petrel_df['id']].copy()
    all_df_petrel['petrel'] = 1
    df = pd.concat([all_df_noise, all_df_petrel])
    df['filename'] = os.path.basename(stem)
    dfs.append(df)
    
df = pd.concat(dfs)

df = df.drop(columns=['freq_peak.2', 'freq_peak.3', 'id'])
df = df.rename(columns={'freq_peak.1': 'freq_peak'})

df.to_csv('../data/features/samples_all.csv', index=False)
