==> notebooks/01_plot_features.ipynb <==
# Feature analysis for storm petrel recordings from St Helena

In this notebook we're going to look at distribution of numerical features we have calculated for selected recordings from St Helena. The purpose is to to determine which seem to best describe storm petrels against *noise*, i.e. any non-petrel sound. 

import os
import sys
import pandas as pd

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
    
import plotly
import plotly.plotly as py
import plotly.graph_objs as go

import audioexplorer.audio_io as aio
import audioexplorer.features as features

Load the data and separate petrels and noise

df = pd.read_csv('s3://stormpetrels/samples/labels/samples_all.csv')
df_petrel = df[df['petrel'] == 1]
df_noise = df[df['petrel'] == 0]

features.FEATURES

Let's use better names so that it's claer what we're computing.

features.FEATURES.update({
    'LPC': 'Linear Predictor Coefficients',
    'LSF': 'Line Spectral Frequency',
    'MFCC': 'Mel-frequencies cepstrum coefficients',
    'OBSI': 'Octave band signal intensity'
})

Let `0` denote *noise* and `1` *petrels*.

df['petrel'].value_counts()

We're going to use `plotly` library to produce boxplots. Follow [getting started guide](https://plot.ly/python/getting-started/) to install and set up the library (free account is needed).

def plot_features(df, selection):
    df_petrel = df[df['petrel'] == 1]
    df_noise = df[df['petrel'] == 0]
    title = features.FEATURES[selection]
    sel_columns = [column_name for column_name in df.columns.values if selection in column_name]
    fig = plotly.tools.make_subplots(rows=len(sel_columns), subplot_titles=tuple(sel_columns))

    marker_petrels = {'color': '#FF4136', 'size': 2}
    marker_noise = {'color': '#0000FF', 'size': 2}

    for idx, name in enumerate(sel_columns):
        trace_petrel = go.Box(x=df_petrel[name], name='petrels', marker=marker_petrels)
        trace_noise = go.Box(x=df_noise[name], name='noise', marker=marker_noise)
        fig.append_trace(trace_petrel, idx+1, 1)
        fig.append_trace(trace_noise, idx+1, 1)

    fig['layout'].update(height=len(sel_columns) * 400, width=1200, title='Storm Petrel: ' + title, showlegend=False)
    return fig

Let's plot them all and put in public space

selection = 'freq'
fig = plot_features(df, selection)
py.plot(fig, filename='Storm Petrels ' + selection)

selection = 'pitch'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'Chroma'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'LPC'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'LSF'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'MFCC'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'OBSI'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'SpectralCrestFactorPerBand'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'SpectralFlatness'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'SpectralFlux'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'SpectralRolloff'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

selection = 'SpectralVariation'
fig = plot_features(df, selection)
py.plot(fig, filename=selection)

nice_features =  ['freq_Q25', 'freq_Q75', 'freq_mean', 'freq_median',
       'freq_mode', 'freq_peak', 'offset', 'onset', 'pitch_IQR',
       'pitch_Q25', 'pitch_Q75', 'pitch_max', 'pitch_mean',
       'pitch_median', 'pitch_min', 'yaafe_LPC', 'yaafe_LSF.0', 'yaafe_LSF.1',
       'yaafe_LSF.2', 'yaafe_LSF.3', 'yaafe_LSF.4', 'yaafe_LSF.5',
       'yaafe_LSF.6', 'yaafe_LSF.7', 'yaafe_LSF.8', 'yaafe_LSF.9',
       'yaafe_MFCC.0', 'yaafe_MFCC.1', 'yaafe_MFCC.10', 'yaafe_MFCC.11',
       'yaafe_MFCC.12', 'yaafe_MFCC.2', 'yaafe_MFCC.3', 'yaafe_MFCC.4',
       'yaafe_MFCC.5', 'yaafe_MFCC.6', 'yaafe_MFCC.7', 'yaafe_MFCC.8',
       'yaafe_MFCC.9', 'yaafe_OBSI.0', 'yaafe_OBSI.1', 'yaafe_OBSI.2',
       'yaafe_OBSI.3', 'yaafe_OBSI.4', 'yaafe_OBSI.5', 'yaafe_OBSI.6',
       'yaafe_OBSI.7', 'yaafe_OBSI.8', 'yaafe_SpectralFlatness',
       'yaafe_SpectralFlux', 'yaafe_SpectralRolloff',
       'yaafe_SpectralVariation', 'petrel', 'filename']

nice_features
